pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = [{ name = "Pedro Gustavo Santana", email = "pedrogustavosantana97@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["django (>=5.1.6,<6.0.0)", "django-rest-framework (>=0.1.0,<0.2.0)", "celery (>=5.5.3,<6.0.0)", "psycopg2 (>=2.9.10,<3.0.0)", "redis (>=6.2.0,<7.0.0)", "python-dotenv (>=1.1.0,<2.0.0)", "django-redis (>=6.0.0,<7.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.production.dependencies]
gunicorn = "^23.0.0"

